Namespace the cache pool adapters
