+++ src/Adapter/Apcu.php
@@ -6,54 +6,30 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-namespace Joomla\Cache;
+namespace Joomla\Cache\Adapter;
 
+use Joomla\Cache\Cache;
 use Joomla\Cache\Item\HasExpirationDateInterface;
 use Joomla\Cache\Item\Item;
 use Psr\Cache\CacheItemInterface;
 
 /**
- * Redis cache driver for the Joomla Framework.
+ * APCu cache driver for the Joomla Framework.
  *
- * @since  1.0
+ * @since  __DEPLOY_VERSION__
  */
-class Redis extends Cache
+class Apcu extends Cache
 {
-	/**
-	 * The redis driver.
-	 *
-	 * @var    \Redis
-	 * @since  1.0
-	 */
-	private $driver;
-
-	/**
-	 * Constructor.
-	 *
-	 * @param   \Redis              $redis    The Redis driver being used for this pool
-	 * @param   array|\ArrayAccess  $options  An options array, or an object that implements \ArrayAccess
-	 *
-	 * @since   __DEPLOY_VERSION__
-	 * @throws  \RuntimeException
-	 */
-	public function __construct(\Redis $redis, $options = [])
-	{
-		// Parent sets up the caching options and checks their type
-		parent::__construct($options);
-
-		$this->driver = $redis;
-	}
-
 	/**
 	 * This will wipe out the entire cache's keys
 	 *
 	 * @return  boolean  The result of the clear operation.
 	 *
-	 * @since   1.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	public function clear()
 	{
-		return $this->driver->flushDB();
+		return apcu_clear_cache();
 	}
 
 	/**
@@ -63,19 +39,52 @@ class Redis extends Cache
 	 *
 	 * @return  CacheItemInterface
 	 *
-	 * @since   1.0
+	 * @since   __DEPLOY_VERSION__
+	 * @throws  \RuntimeException
 	 */
 	public function getItem($key)
 	{
-		$value = $this->driver->get($key);
+		$success = false;
+		$value = apcu_fetch($key, $success);
 		$item = new Item($key);
 
-		if ($value !== false)
+		if ($success)
 		{
 			$item->set($value);
 		}
 
 		return $item;
+	}
+
+	/**
+	 * Obtain multiple CacheItems by their unique keys.
+	 *
+	 * @param   array  $keys  A list of keys that can obtained in a single operation.
+	 *
+	 * @return  array  An associative array of CacheItem objects keyed on the cache key.
+	 *
+	 * @since   __DEPLOY_VERSION__
+	 */
+	public function getItems(array $keys = array())
+	{
+		$items = array();
+		$success = false;
+		$values = apcu_fetch($keys, $success);
+
+		if ($success && is_array($values))
+		{
+			foreach ($keys as $key)
+			{
+				$items[$key] = new Item($key);
+
+				if (isset($values[$key]))
+				{
+					$items[$key]->set($values[$key]);
+				}
+			}
+		}
+
+		return $items;
 	}
 
 	/**
@@ -85,13 +94,13 @@ class Redis extends Cache
 	 *
 	 * @return  boolean
 	 *
-	 * @since   1.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	public function deleteItem($key)
 	{
 		if ($this->hasItem($key))
 		{
-			return (bool) $this->driver->del($key);
+			return apcu_delete($key);
 		}
 
 		// If the item doesn't exist, no error
@@ -103,21 +112,22 @@ class Redis extends Cache
 	 *
 	 * @param   CacheItemInterface  $item  The cache item to save.
 	 *
-	 * @return  bool  True if the item was successfully persisted. False if there was an error.
+	 * @return static
+	 *   The invoked object.
 	 */
 	public function save(CacheItemInterface $item)
 	{
+		// If we are able to find out when the item expires - find out. Else bail.
 		if ($item instanceof HasExpirationDateInterface)
 		{
 			$ttl = $this->convertItemExpiryToSeconds($item);
-
-			if ($ttl > 0)
-			{
-				return $this->driver->setex($item->getKey(), $ttl, $item->get());
-			}
+		}
+		else
+		{
+			$ttl = 0;
 		}
 
-		return $this->driver->set($item->getKey(), $item->get());
+		return apcu_store($item->getKey(), $item->get(), $ttl);
 	}
 
 	/**
@@ -127,11 +137,11 @@ class Redis extends Cache
 	 *
 	 * @return  boolean
 	 *
-	 * @since   1.0
+	 * @since   __DEPLOY_VERSION__
 	 */
 	public function hasItem($key)
 	{
-		return $this->driver->exists($key);
+		return apcu_exists($key);
 	}
 
 	/**
@@ -143,6 +153,14 @@ class Redis extends Cache
 	 */
 	public static function isSupported()
 	{
-		return (extension_loaded('redis') && class_exists('Redis'));
+		$supported = extension_loaded('apcu') && ini_get('apc.enabled');
+
+		// If on the CLI interface, the `apc.enable_cli` option must also be enabled
+		if ($supported && php_sapi_name() === 'cli')
+		{
+			$supported = ini_get('apc.enable_cli');
+		}
+
+		return (bool) $supported;
 	}
 }
