+++ src/Adapter/File.php
@@ -6,13 +6,14 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-namespace Joomla\Cache;
+namespace Joomla\Cache\Adapter;
 
+use Joomla\Cache\Cache;
+use Joomla\Cache\Exception\InvalidArgumentException;
 use Joomla\Cache\Exception\RuntimeException;
 use Joomla\Cache\Item\HasExpirationDateInterface;
-use Psr\Cache\CacheItemInterface;
-use Joomla\Cache\Exception\InvalidArgumentException;
 use Joomla\Cache\Item\Item;
+use Psr\Cache\CacheItemInterface;
 
 /**
  * Filesystem cache driver for the Joomla Framework.
