+++ src/Adapter/None.php
@@ -6,10 +6,11 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-namespace Joomla\Cache;
+namespace Joomla\Cache\Adapter;
 
-use Psr\Cache\CacheItemInterface;
+use Joomla\Cache\Cache;
 use Joomla\Cache\Item\Item;
+use Psr\Cache\CacheItemInterface;
 
 /**
  * Runtime cache only driver for the Joomla Framework.
