+++ tests/Adapter/XCacheTest.php
@@ -4,14 +4,15 @@
  * @license    GNU General Public License version 2 or later; see LICENSE
  */
 
-namespace Joomla\Cache\Tests;
+namespace Joomla\Cache\Tests\Adapter;
 
-use Joomla\Cache;
+use Joomla\Cache\Adapter\XCache;
+use Joomla\Cache\Tests\CacheTest;
 
 /**
- * Tests for the Joomla\Cache\Wincache class.
+ * Tests for the Joomla\Cache\Adapter\XCache class.
  */
-class WincacheTest extends CacheTest
+class XCacheTest extends CacheTest
 {
 	/**
 	 * Sets up the fixture, for example, open a network connection.
@@ -21,11 +22,11 @@ class WincacheTest extends CacheTest
 	{
 		parent::setUp();
 
-		if (!Cache\Wincache::isSupported())
+		if (!XCache::isSupported())
 		{
-			$this->markTestSkipped('WinCache Cache Handler is not supported on this system.');
+			$this->markTestSkipped('XCache Cache Handler is not supported on this system.');
 		}
 
-		$this->instance = new Cache\Wincache($this->cacheOptions);
+		$this->instance = new XCache($this->cacheOptions);
 	}
 }
